requirements.txt
streamlit>=1.20.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
joblib>=1.2.0
Pillow>=9.5.0